+++ farmbot_core/lib/farmbot_celery_script/compilers/axis_control_compiler.ex
@@ -2,23 +2,17 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
   alias FarmbotCeleryScript.Compiler
 
   # Compiles move_absolute
-  def move_absolute(
-        %{args: %{location: location, offset: offset, speed: speed}},
-        env
-      ) do
+  def move_absolute(%{args: %{location: location,offset: offset,speed: speed}}, cs_scope) do
+    [locx , locy , locz] = cs_to_xyz(location, cs_scope)
+    [offx, offy, offz] = cs_to_xyz(offset, cs_scope)
+
     quote location: :keep do
-      # Extract the location arg
-      with %{x: locx, y: locy, z: locz} =
-             unquote(Compiler.compile_ast(location, env)),
-           # Extract the offset arg
-           %{x: offx, y: offy, z: offz} =
-             unquote(Compiler.compile_ast(offset, env)) do
         # Subtract the location from offset.
         # Note: list syntax here for readability.
         [x, y, z] = [
-          locx + offx,
-          locy + offy,
-          locz + offz
+          unquote(locx) + unquote(offx),
+          unquote(locy) + unquote(offy),
+          unquote(locz) + unquote(offz)
         ]
 
         x_str = FarmbotCeleryScript.FormatUtil.format_float(x)
@@ -34,18 +28,17 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
           x,
           y,
           z,
-          unquote(Compiler.compile_ast(speed, env))
+          unquote(Compiler.celery_to_elixir(speed, cs_scope))
         )
-      end
     end
   end
 
   # compiles move_relative into move absolute
-  def move_relative(%{args: %{x: x, y: y, z: z, speed: speed}}, env) do
+  def move_relative(%{args: %{x: x, y: y, z: z, speed: speed}}, cs_scope) do
     quote location: :keep do
-      with locx when is_number(locx) <- unquote(Compiler.compile_ast(x, env)),
-           locy when is_number(locy) <- unquote(Compiler.compile_ast(y, env)),
-           locz when is_number(locz) <- unquote(Compiler.compile_ast(z, env)),
+      with locx when is_number(locx) <- unquote(Compiler.celery_to_elixir(x, cs_scope)),
+           locy when is_number(locy) <- unquote(Compiler.celery_to_elixir(y, cs_scope)),
+           locz when is_number(locz) <- unquote(Compiler.celery_to_elixir(z, cs_scope)),
            curx when is_number(curx) <-
              FarmbotCeleryScript.SysCalls.get_current_x(),
            cury when is_number(cury) <-
@@ -69,14 +62,14 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
           x,
           y,
           z,
-          unquote(Compiler.compile_ast(speed, env))
+          unquote(Compiler.celery_to_elixir(speed, cs_scope))
         )
       end
     end
   end
 
   # Expands find_home(all) into three find_home/1 calls
-  def find_home(%{args: %{axis: "all"}}, _env) do
+  def find_home(%{args: %{axis: "all"}}, _cs_scope) do
     quote location: :keep do
       FarmbotCeleryScript.SysCalls.log("Finding home on all axes", true)
 
@@ -88,10 +81,10 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
   end
 
   # compiles find_home
-  def find_home(%{args: %{axis: axis}}, env) do
+  def find_home(%{args: %{axis: axis}}, cs_scope) do
     quote location: :keep do
       with axis when axis in ["x", "y", "z"] <-
-             unquote(Compiler.compile_ast(axis, env)) do
+             unquote(Compiler.celery_to_elixir(axis, cs_scope)) do
         FarmbotCeleryScript.SysCalls.log(
           "Finding home on the #{String.upcase(axis)} axis",
           true
@@ -106,12 +99,12 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
   end
 
   # Expands home(all) into three home/1 calls
-  def home(%{args: %{axis: "all", speed: speed}}, env) do
+  def home(%{args: %{axis: "all", speed: speed}}, cs_scope) do
     quote location: :keep do
       FarmbotCeleryScript.SysCalls.log("Going to home on all axes", true)
 
       with speed when is_number(speed) <-
-             unquote(Compiler.compile_ast(speed, env)),
+             unquote(Compiler.celery_to_elixir(speed, cs_scope)),
            :ok <- FarmbotCeleryScript.SysCalls.home("z", speed),
            :ok <- FarmbotCeleryScript.SysCalls.home("y", speed) do
         FarmbotCeleryScript.SysCalls.home("x", speed)
@@ -120,12 +113,12 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
   end
 
   # compiles home
-  def home(%{args: %{axis: axis, speed: speed}}, env) do
+  def home(%{args: %{axis: axis, speed: speed}}, cs_scope) do
     quote location: :keep do
       with axis when axis in ["x", "y", "z"] <-
-             unquote(Compiler.compile_ast(axis, env)),
+             unquote(Compiler.celery_to_elixir(axis, cs_scope)),
            speed when is_number(speed) <-
-             unquote(Compiler.compile_ast(speed, env)) do
+             unquote(Compiler.celery_to_elixir(speed, cs_scope)) do
         FarmbotCeleryScript.SysCalls.log(
           "Going to home on the #{String.upcase(axis)} axis",
           true
@@ -140,7 +133,7 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
   end
 
   # Expands zero(all) into three zero/1 calls
-  def zero(%{args: %{axis: "all"}}, _env) do
+  def zero(%{args: %{axis: "all"}}, _cs_scope) do
     quote location: :keep do
       FarmbotCeleryScript.SysCalls.log("Setting home for all axes", true)
 
@@ -152,10 +145,10 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
   end
 
   # compiles zero
-  def zero(%{args: %{axis: axis}}, env) do
+  def zero(%{args: %{axis: axis}}, cs_scope) do
     quote location: :keep do
       with axis when axis in ["x", "y", "z"] <-
-             unquote(Compiler.compile_ast(axis, env)) do
+             unquote(Compiler.celery_to_elixir(axis, cs_scope)) do
         FarmbotCeleryScript.SysCalls.log(
           "Setting home for the #{String.upcase(axis)} axis",
           true
@@ -170,7 +163,7 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
   end
 
   # Expands calibrate(all) into three calibrate/1 calls
-  def calibrate(%{args: %{axis: "all"}}, _env) do
+  def calibrate(%{args: %{axis: "all"}}, _cs_scope) do
     quote location: :keep do
       FarmbotCeleryScript.SysCalls.log("Finding length of all axes", true)
 
@@ -184,10 +177,10 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
   end
 
   # compiles calibrate
-  def calibrate(%{args: %{axis: axis}}, env) do
+  def calibrate(%{args: %{axis: axis}}, cs_scope) do
     quote location: :keep do
       with axis when axis in ["x", "y", "z"] <-
-             unquote(Compiler.compile_ast(axis, env)) do
+             unquote(Compiler.celery_to_elixir(axis, cs_scope)) do
         msg = "Determining length of the #{String.upcase(axis)} axis"
         FarmbotCeleryScript.SysCalls.log(msg, true)
         FarmbotCeleryScript.SysCalls.calibrate(axis)
@@ -196,4 +189,32 @@ defmodule FarmbotCeleryScript.Compiler.AxisControl do
       end
     end
   end
+
+  defp cs_to_xyz(%{kind: :identifier} = ast, cs_scope) do
+    label = ast.args.label
+    {:ok, variable} = FarmbotCeleryScript.Compiler.Scope.fetch!(cs_scope, label)
+    # Prevent circular refernces.
+    # I doubt end users would intentionally do this, so treat
+    # it like an error.
+    if variable.kind == :identifier, do: raise "Refusing to perform recursion"
+    cs_to_xyz(variable, cs_scope)
+  end
+
+  defp cs_to_xyz(%{kind: :coordinate} = ast, _) do
+    vec_map_to_array(ast.args)
+  end
+
+  defp cs_to_xyz(%{kind: :tool, args: args}, _) do
+    slot = FarmbotCeleryScript.SysCalls.get_toolslot_for_tool(args.tool_id)
+    vec_map_to_array(slot)
+  end
+
+  defp cs_to_xyz(%{kind: :point} = ast, _) do
+    %{ pointer_type: t, pointer_id: id } = ast.args
+    vec_map_to_array(FarmbotCeleryScript.SysCalls.point(t, id))
+  end
+
+  defp cs_to_xyz(other, _), do: raise "Unexpected location or offset: #{inspect(other)}"
+
+  defp vec_map_to_array(xyz), do: [ xyz.x, xyz.y, xyz.z ]
 end