+++ farmbot_core/lib/farmbot_celery_script/compilers/lua.ex
@@ -1,22 +1,24 @@
 defmodule FarmbotCeleryScript.Compiler.Lua do
   alias FarmbotCeleryScript.SysCalls
-  alias FarmbotCeleryScript.Compiler.VariableTransformer
+  alias FarmbotCeleryScript.Compiler.{ VariableTransformer, Scope, }
 
-  def lua(%{args: %{lua: lua}}, _env) do
+  def lua(%{args: %{lua: lua}}, cs_scope) do
     quote location: :keep do
+      # lua.ex
       mod = unquote(__MODULE__)
-      mod.do_lua(unquote(lua), better_params)
+      mod.do_lua(unquote(lua), unquote(cs_scope))
     end
   end
 
-  def do_lua(lua, better_params) do
-    go = fn params, label, lua ->
-      {VariableTransformer.run!(params[label]), lua}
+  def do_lua(lua, cs_scope) do
+    go = fn label, lua ->
+      {:ok, variable} = Scope.fetch!(cs_scope, label)
+      {VariableTransformer.run!(variable), lua}
     end
 
     lookup = fn
-      [label], lua -> go.(better_params, label, lua)
-      [], lua -> go.(better_params, "parent", lua)
+      [label], lua -> go.(label, lua)
+      [], lua -> go.("parent", lua)
       _, _ -> %{error: "Invalid input. Please pass 1 variable name (string)."}
     end