+++ farmbot_core/test/farmware_runtime_test.exs
@@ -3,6 +3,19 @@ defmodule FarmbotCore.FarmwareRuntimeTest do
   import ExUnit.CaptureLog
   alias FarmbotCore.FarmwareRuntime
 
+  test "error handling" do
+    state = %FarmwareRuntime{
+      scheduler_ref: make_ref(),
+      caller: self()
+    }
+
+    error = {:error, "intentional error (unit test)"}
+    message = {:csvm_done, state.scheduler_ref, error}
+    {:noreply, next_state} = FarmwareRuntime.handle_info(message, state)
+    refute next_state.scheduler_ref
+    assert_receive ^error, 5000
+  end
+
   test "stop/1" do
     fake_pid = spawn(fn -> 2 + 2 end)
     stop = fn -> FarmwareRuntime.stop(fake_pid) end